refactor(tests): make test_copy and async counterpart more similar
