+++ tests/crdb/test_copy_async.py
@@ -8,9 +8,8 @@ from psycopg.adapt import PyFormat
 from psycopg.types.numeric import Int4
 
 from ..utils import eur, gc_collect, gc_count
-from ..test_copy import sample_text, sample_binary  # noqa
-from ..test_copy import sample_records
-from ..test_copy_async import ensure_table
+from .._test_copy import sample_text, sample_binary  # noqa
+from .._test_copy import ensure_table_async, sample_records
 from .test_copy import sample_tabledef, copyopt
 
 pytestmark = [pytest.mark.crdb, pytest.mark.anyio]
@@ -22,7 +21,7 @@ pytestmark = [pytest.mark.crdb, pytest.mark.anyio]
 )
 async def test_copy_in_buffers(aconn, format, buffer):
     cur = aconn.cursor()
-    await ensure_table(cur, sample_tabledef)
+    await ensure_table_async(cur, sample_tabledef)
     async with cur.copy(f"copy copy_in from stdin {copyopt(format)}") as copy:
         await copy.write(globals()[buffer])
 
@@ -33,7 +32,7 @@ async def test_copy_in_buffers(aconn, format, buffer):
 
 async def test_copy_in_buffers_pg_error(aconn):
     cur = aconn.cursor()
-    await ensure_table(cur, sample_tabledef)
+    await ensure_table_async(cur, sample_tabledef)
     with pytest.raises(e.UniqueViolation):
         async with cur.copy("copy copy_in from stdin") as copy:
             await copy.write(sample_text)
@@ -43,7 +42,7 @@ async def test_copy_in_buffers_pg_error(aconn):
 
 async def test_copy_in_str(aconn):
     cur = aconn.cursor()
-    await ensure_table(cur, sample_tabledef)
+    await ensure_table_async(cur, sample_tabledef)
     async with cur.copy("copy copy_in from stdin") as copy:
         await copy.write(sample_text.decode())
 
@@ -55,7 +54,7 @@ async def test_copy_in_str(aconn):
 @pytest.mark.xfail(reason="bad sqlstate - CRDB #81559")
 async def test_copy_in_error(aconn):
     cur = aconn.cursor()
-    await ensure_table(cur, sample_tabledef)
+    await ensure_table_async(cur, sample_tabledef)
     with pytest.raises(e.QueryCanceled):
         async with cur.copy("copy copy_in from stdin with binary") as copy:
             await copy.write(sample_text.decode())
@@ -66,7 +65,7 @@ async def test_copy_in_error(aconn):
 @pytest.mark.parametrize("format", Format)
 async def test_copy_in_empty(aconn, format):
     cur = aconn.cursor()
-    await ensure_table(cur, sample_tabledef)
+    await ensure_table_async(cur, sample_tabledef)
     async with cur.copy(f"copy copy_in from stdin {copyopt(format)}"):
         pass
 
@@ -77,7 +76,7 @@ async def test_copy_in_empty(aconn, format):
 @pytest.mark.slow
 async def test_copy_big_size_record(aconn):
     cur = aconn.cursor()
-    await ensure_table(cur, "id serial primary key, data text")
+    await ensure_table_async(cur, "id serial primary key, data text")
     data = "".join(chr(randrange(1, 256)) for i in range(10 * 1024 * 1024))
     async with cur.copy("copy copy_in (data) from stdin") as copy:
         await copy.write_row([data])
@@ -89,7 +88,7 @@ async def test_copy_big_size_record(aconn):
 @pytest.mark.slow
 async def test_copy_big_size_block(aconn):
     cur = aconn.cursor()
-    await ensure_table(cur, "id serial primary key, data text")
+    await ensure_table_async(cur, "id serial primary key, data text")
     data = "".join(choice(string.ascii_letters) for i in range(10 * 1024 * 1024))
     copy_data = data + "\n"
     async with cur.copy("copy copy_in (data) from stdin") as copy:
@@ -101,7 +100,7 @@ async def test_copy_big_size_block(aconn):
 
 async def test_copy_in_buffers_with_pg_error(aconn):
     cur = aconn.cursor()
-    await ensure_table(cur, sample_tabledef)
+    await ensure_table_async(cur, sample_tabledef)
     with pytest.raises(e.UniqueViolation):
         async with cur.copy("copy copy_in from stdin") as copy:
             await copy.write(sample_text)
@@ -113,7 +112,7 @@ async def test_copy_in_buffers_with_pg_error(aconn):
 @pytest.mark.parametrize("format", Format)
 async def test_copy_in_records(aconn, format):
     cur = aconn.cursor()
-    await ensure_table(cur, sample_tabledef)
+    await ensure_table_async(cur, sample_tabledef)
 
     async with cur.copy(f"copy copy_in from stdin {copyopt(format)}") as copy:
         for row in sample_records:
@@ -131,7 +130,7 @@ async def test_copy_in_records(aconn, format):
 @pytest.mark.parametrize("format", Format)
 async def test_copy_in_records_set_types(aconn, format):
     cur = aconn.cursor()
-    await ensure_table(cur, sample_tabledef)
+    await ensure_table_async(cur, sample_tabledef)
 
     async with cur.copy(f"copy copy_in from stdin {copyopt(format)}") as copy:
         copy.set_types(["int4", "int4", "text"])
@@ -146,7 +145,7 @@ async def test_copy_in_records_set_types(aconn, format):
 @pytest.mark.parametrize("format", Format)
 async def test_copy_in_records_binary(aconn, format):
     cur = aconn.cursor()
-    await ensure_table(cur, "col1 serial primary key, col2 int4, data text")
+    await ensure_table_async(cur, "col1 serial primary key, col2 int4, data text")
 
     async with cur.copy(
         f"copy copy_in (col2, data) from stdin {copyopt(format)}"
@@ -162,7 +161,7 @@ async def test_copy_in_records_binary(aconn, format):
 @pytest.mark.crdb_skip("copy canceled")
 async def test_copy_in_buffers_with_py_error(aconn):
     cur = aconn.cursor()
-    await ensure_table(cur, sample_tabledef)
+    await ensure_table_async(cur, sample_tabledef)
     with pytest.raises(e.QueryCanceled) as exc:
         async with cur.copy("copy copy_in from stdin") as copy:
             await copy.write(sample_text)
@@ -174,7 +173,7 @@ async def test_copy_in_buffers_with_py_error(aconn):
 
 async def test_copy_in_allchars(aconn):
     cur = aconn.cursor()
-    await ensure_table(cur, "col1 int primary key, col2 int, data text")
+    await ensure_table_async(cur, "col1 int primary key, col2 int, data text")
 
     async with cur.copy("copy copy_in from stdin") as copy:
         for i in range(1, 256):
